chore: refactor Editor
